Analyza/PDF Parser/requirements.txt
boto3==1.24.7
botocore==1.27.7
certifi==2022.5.18.1
charset-normalizer==2.0.12
ci-info==0.2.0
click==8.1.3
colorama==0.4.4
configobj==5.0.6
configparser==5.2.0
etelemetry==0.3.0
filelock==3.7.1
future==0.18.2
httplib2==0.20.4
idna==3.3
isodate==0.6.1
jmespath==1.0.0
looseversion==1.0.1
lxml==4.9.0
mysql-connector-python==8.0.29
networkx==2.8.3
nibabel==3.2.2
nipype==1.8.2
numpy==1.22.4
packaging==21.3
pandas==1.4.2
pathlib==1.0.1
protobuf==4.21.1
prov==2.0.0
pydot==1.4.2
PyMuPDF==1.19.6
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
pyxnat==1.4
rdflib==6.1.1
requests==2.28.0
s3transfer==0.6.0
scipy==1.8.1
simplejson==3.17.6
six==1.16.0
traits==6.3.2
urllib3==1.26.9
